% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emulateAPSIM.R
\name{apsim_emul_sa}
\alias{apsim_emul_sa}
\title{A function to estimate sensitivity indices based on emulation methods}
\usage{
apsim_emul_sa(model, X, y = NULL, method, ...)
}
\arguments{
\item{model}{The model that is of interest}

\item{X}{The matrix of inputs at which to evaluate the model}

\item{y}{Optional vector of model evaluations that can be used in place of the model statement}

\item{method}{The method to use to emulate the model}

\item{...}{Additional arguments passed to the choosen method}
}
\value{
A data frame of results, the exact form depends on the method
}
\description{
This is a generic function that can be used to estimate the sensitivity
indices for complex computer models using GAM-based emulators.
}
\examples{
\dontrun{
meanYield<-function(x){
 return(mean(x$lai_cowpea))
}
n <- 75
parValues <- data.frame(SoilCN = runif(n, 5, 25),
                 DiffusConst = runif(n, 20, 50), CNCov = runif(n, 0, 1))
apsimExe <-"C:/Program Files (x86)/Apsim75-r3008/Model/Apsim.exe"
apsimWd <- "~/APSIM"
apsimVars <- c("SoilOrganicMatter/SoilCN", "SoilWater/DiffusConst", "SoilWater/CNCov")
apsimFile <- "Canopy.apsim"
emulRes <- apsim_emul_sa(model = apsim_vector, X = parValues, method = "singleGAM", exe = apsimExe, wd = apsimWd,
                          vars = apsimVars, to.run = apsimFile, g = meanYield, overwrite = TRUE)
plot(emulRes)

}
}

