% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emulateAPSIM.R
\name{apsim_emulator}
\alias{apsim_emulator}
\title{A function to estimate sensitivity indices based on emulation methods}
\usage{
apsim_emulator(model, X, y = NULL, method, ...)
}
\arguments{
\item{model}{The model that is of interest}

\item{X}{The matrix of inputs at which to evaluate the model}

\item{y}{Optional vector of model evaluations that can be used in place of the model statement}

\item{method}{The method to use to emulate the model}

\item{...}{Additional arguments passed to the choosen method}
}
\value{
A data frame of results, the exact form depends on the method
}
\description{
This is a generic function that can be used to estimate the sensitivity
indices for complex computer models using GAM-based emulators.
}

