% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstention.R
\name{is_apsim_abstention}
\alias{is_apsim_abstention}
\title{Is an object an apsimR abstention?}
\usage{
is_apsim_abstention(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A single logical: \code{TRUE} when \code{x} is an \code{apsim_abstention}.
}
\description{
Is an object an apsimR abstention?
}
\examples{
is_apsim_abstention(apsim_abstention("run_failed"))
is_apsim_abstention(42)
}
