% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emulateAPSIM.R
\name{plot.gamSA}
\alias{plot.gamSA}
\title{Plot Sensitivity Analysis Results}
\usage{
\method{plot}{gamSA}(x, ...)
}
\arguments{
\item{x,...}{Results of a successful call to \code{apsim_emulator}}
}
\value{
The results of the sensitivity analysis without the estimated output and residuals
}
\description{
Plot routine for \code{"gamSA"} class objects
}
\details{
The default plot method for the results of a sensitivity analysis of APSIM using
the single or separate GAM-based emulator.  This function produces a bar chart where the heights of the
bar represents the magnitude of the estimated sensitivity index estimate.  Error bars represent
bootstrap calibrated confidence regions.
}

