% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emulateAPSIM.R
\name{print.gamSA}
\alias{print.gamSA}
\title{Print Sensitivity Analysis Results}
\usage{
\method{print}{gamSA}(x, ...)
}
\arguments{
\item{x,...}{Results of a successful call to \code{apsim_emulator}}
}
\value{
The results of the sensitivity analysis without the estimated output and residuals
}
\description{
The default print method for the results of a sensitivity analysis of APSIM using
the single or separate GAM-based emulator.
}

